Refactor swtbot based tests
